# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RunAPP API",
        "description": "API dokumentacija za RunAPP aplikaciju",
        "contact": {
            "name": "API Support",
            "url": "https://example.com/support",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/trke": {
            "get": {
                "tags": [
                    "Trka"
                ],
                "summary": "Get lista trka",
                "description": "Vraca listu svih trka",
                "operationId": "3de17ccb07a34e1b96acf45eb5a357ee",
                "responses": {
                    "200": {
                        "description": "Uspesno vracene trke",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Trka"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Trka"
                ],
                "summary": "Kreiranje trke",
                "description": "Kreira novu trku",
                "operationId": "20d2c39e36cbc5399fe6b9552ce2d533",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "naziv",
                                    "datum",
                                    "lokacija_id",
                                    "godina",
                                    "organizator",
                                    "kilometraza"
                                ],
                                "properties": {
                                    "naziv": {
                                        "type": "string",
                                        "example": "Beogradski maraton"
                                    },
                                    "datum": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-05-15"
                                    },
                                    "lokacija_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "godina": {
                                        "type": "integer",
                                        "example": 2026
                                    },
                                    "organizator": {
                                        "type": "string",
                                        "example": "Atletski savez"
                                    },
                                    "kilometraza": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 42.195
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Trka uspesno kreirana",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Trka"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validaciona greska"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/trke/{id}": {
            "get": {
                "tags": [
                    "Trka"
                ],
                "summary": "Get trka po ID-u",
                "description": "Vraca detalje jedne trke",
                "operationId": "2944fb2ae9886c7515fa6291b188abf0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID trke",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspesno vracena trka",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Trka"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Trka nije pronadjena"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Trka"
                ],
                "summary": "Brisanje trke",
                "description": "Brise trku (samo admin)",
                "operationId": "385f5bef7b00194aac3e973ad2dae177",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID trke",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trka uspesno obrisana"
                    },
                    "404": {
                        "description": "Trka nije pronadjena"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    },
                    "403": {
                        "description": "Zabranjen pristup"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/trke/{id}/ucesca": {
            "get": {
                "tags": [
                    "Trka"
                ],
                "summary": "Get ucesca za trku",
                "description": "Vraca sva ucesca za odredjenu trku",
                "operationId": "ed76e2fe5210f9dc0acb3eb4c848ecd8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID trke",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspesno vracena ucesca za trku",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Ucesce"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Trka nije pronadjena"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/trke/{trkaId}/ucesca": {
            "get": {
                "tags": [
                    "Ucesce"
                ],
                "summary": "Get ucesca po trci",
                "description": "Vraca sva ucesca za odredjenu trku",
                "operationId": "c9174f857551e4f6ad61e0a4648f296b",
                "parameters": [
                    {
                        "name": "trkaId",
                        "in": "path",
                        "description": "ID trke",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspesno vracena ucesca za trku",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Ucesce"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Trka": {
                "title": "Trka",
                "description": "Trka model",
                "required": [
                    "id",
                    "naziv",
                    "godina",
                    "organizator",
                    "kilometraza",
                    "datum",
                    "lokacija_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "naziv": {
                        "type": "string",
                        "example": "Maraton Beograd"
                    },
                    "godina": {
                        "type": "integer",
                        "example": 2024
                    },
                    "organizator": {
                        "type": "string",
                        "example": "Atletski savez Srbije"
                    },
                    "kilometraza": {
                        "type": "number",
                        "format": "float",
                        "example": 42.195
                    },
                    "datum": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-01T09:00:00Z"
                    },
                    "lokacija_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
            "Ucesce": {
                "title": "Ucesce",
                "description": "Ucesce model",
                "required": [
                    "id",
                    "trka_id",
                    "user_id",
                    "vreme"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "trka_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "vreme": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}